Added a file to merge

+++ lyon_merge.py
@@ -18,8 +18,23 @@ if not os.path.exists(lyon_data_file02):
     exit(1)
 # Read the file
 rawData02 = pd.read_csv(lyon_data_file02)
+# Read data/lyon_data03.csv
+lyon_data_file03 = 'data/lyon_data03.csv'
+if not os.path.exists(lyon_data_file03):
+    print(f"xxx File {lyon_data_file03} does not exist.")
+    exit(1)
+# Read the file
+rawData03 = pd.read_csv(lyon_data_file03)
+# Read data/lyon_data04.csv
+lyon_data_file04 = 'data/lyon_data04.csv'
+if not os.path.exists(lyon_data_file04):
+    print(f"xxx File {lyon_data_file04} does not exist.")
+    exit(1)
+# Read the file
+rawData04 = pd.read_csv(lyon_data_file04)
+
 # Concatenate the two dataframes
-rawDataAll = pd.concat([rawData01, rawData02], ignore_index=True)
+rawDataAll = pd.concat([rawData01, rawData02, rawData03, rawData04], ignore_index=True)
 # Save into data/lyon_data.csv
 lyon_data_file = 'data/lyon_data.csv'
 rawDataAll.to_csv(lyon_data_file, index=False)
